refactor: move the item inventory amount from inventory to its own Item() class
Each item in the game is now an instance of the Item() class and subclasses. The class will contain the item's description, inventory amount, and additional attibutes of the item depending on the item type.

A new dictionary called 'items' matches the string of each item to the dictionary value of their respective Item() class instances.

The codes regarding inventory amounts in inventory.py and craft.py were changed in order to incorporate the new dictionary and classes.